Disallow 64-bit int fromarray(); tests force arrays to 32-bit floats.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2929 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/tests/python/test.py
@@ -750,7 +750,7 @@ class AreaTests(OpenCVTests):
 
             def convert(numpydims):
                 """ Create a numpy array with specified dims, return the OpenCV CvMat """
-                a1 = numpy.array([1] * reduce(operator.__mul__, numpydims)).reshape(*numpydims)
+                a1 = numpy.array([1] * reduce(operator.__mul__, numpydims)).reshape(*numpydims).astype(numpy.float32)
                 return cv.fromarray(a1)
             def row_col_chan(m):
                 col = m.cols
